REDUNDANCY TESTS: Tests on foreach and plyr examples, no reports on timing.
We also exclude a 4 inefficient examples that can take a very long
time to process using batch*_local backends.  The reason for this is that
the divide-and-conquer for parallelization is not done where it is most
efficient to do so.  By excluding this, we can now test 89 dplyr examples
on all future strategies in a few minutes.